fix: fix preset resolution when resolved with npx
There are two issues in play here:

1. `require.resolve(..)` returns the path to file referenced by `main` field in
`package.json` while `resolveUsingNpx(..)` return the path to the directory.

2. `resolveUsingNpx` incorrectly assumes the installation folder will be in
`lib` but this assumption breaks when it is installed into `lib64`.

This quick-and-dirty solution tries both `lib` and `lib64`. I am not aware of
any API to get the proper path since I'm not familiar with npx internals.