fix(test): add `NPM_CONFIG_CACHE` dir support
The 'should install an npm module transparently' test did not account
for the possibility that the `NPM_CONFIG_CACHE` environment variable
might be set. So, it would fail in that case. The fix includes checking
the value of `NPM_CONFIG_CACHE` and using that value if it is set,
otherwise defaulting to `.npm` directory.